Twitter_auto_uploader/requirements.txt
tweepy==3.11.0
Pillow==8.4.0
tkcalendar==1.6.1

